pyproject.toml
[project]
name = "cmov"
version = "0.1.0"
description = "A high-level Python library for programmatic timeline-based video animation and dynamic scene creation."
authors = [
    {name = "Dylan DiBeneditto", email = "dibenedittod@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow",
    "fmov"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools]
packages = {find = {where = ["src"]}}
